code refactor for components
