+++ job-listing-page/src/components/LogoItem/LogoItem.css
