+++ job-listing-page/src/components/Subtitle/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../Subtitle.css';
+import './Subtitle.css';
 
 const Subtitle = () => {
     return (
